Added the add-on package
This is what Google Gemini+GitHub CoPilot helped me create.Feel free to download, edit and rebuild the source should you wish to.
BTW, if it Does End Up Working for you on NVDA2025.1, Do Push It Here so I Can See the Update(s) done to it?
Many massive thanks in advance.